Python port of Nailgun client
We have a lot of Buck users who struggle to compile the native Nailgun C client.

This ports the client to pure Python instead. I tested this with the Buck
Nailgun server on Linux 3.10.75 with Python 2.6 and OS X 10.10.4 with Python 2.7.

This client doesn't yet work on Windows, since Python `select.select()` doesn't
handle a mix of sockets and file handles on Windows.