Use java.util.logging for logging (instead of println)
This improves Nailgun usability on dedicated servers. The administrator can
provide the application with a logging.properties file to override its logging
behavior. Log messages can be redirected to files or servers, their formatting
can be customized and so on.
The default logging.properties file can be found in the jre/lib directory.
To provide a custom file, set the java.util.logging.config.file JVM system
property.