[python3] Ensure UTF8 strings are sent properly
Previously we'd take a `buf` (the data to send), use that length in the
framing that we sent (`<ic` in sendall), and then send a binary buffer.

This caused problems when people had UTF8 strings in their environment
under python3, as `len(buf)` is the number of chars, and `len(bbuf)` is
the number of bytes. Multi-byte characters thus sent more data than the
server was expect. This was not a problem in python2, as os.environ just
reads these as binary strings and len() returns the number of bytes.

Test Plan:
Added a new unittest that failed under python3 previously, and passes
now under python2 + python3