[python3] Remove bytes -> str conversion
This was incorrectly done during the python3 migration. stdout/err only
accepts strings in python3, so we were converting to utf-8, which is not
always the right behavior. Instead write to std*.buffer, and flush after
each write.

Tested in buck with python3, and it handles non-utf8 coming over ng now
where it didn't before